+++ cli/tests/test_git_import_export.rs
@@ -58,8 +58,10 @@ fn test_resolution_of_git_tracking_bookmarks() {
     a7f9930bb6d54ba39e6c254135b9bfe32041fea4 old_message
     [EOF]
     ");
-    // Can't be selected by remote_bookmarks()
-    insta::assert_snapshot!(query(r#"remote_bookmarks(exact:"main", exact:"git")"#), @"");
+    insta::assert_snapshot!(query(r#"remote_bookmarks(exact:"main", exact:"git")"#), @r"
+    a7f9930bb6d54ba39e6c254135b9bfe32041fea4 old_message
+    [EOF]
+    ");
 }
 
 #[test]